requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
apscheduler
lxml
cryptography
zeep
python-multipart
